+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.18.15 - 3.11.2025
 - fix OJP1.0 TR sharing - [OJP 1.0 - fix multimodal sharing #300](https://github.com/openTdataCH/ojp-demo-app-src/issues/300) - [PR #208](https://github.com/openTdataCH/ojp-js/pull/208)
+- fix Infeasible typo - [Typo in trip status "Infeasable" #292](https://github.com/openTdataCH/ojp-demo-app-src/issues/292) - [PR #209](https://github.com/openTdataCH/ojp-js/pull/209)
 
 ## 0.18.14 - 24.10.2025
 - dont add IndividualTransportOption if no custom mode or additional properties are set - [PR #204](https://github.com/openTdataCH/ojp-js/pull/204)